refactor: restructure Results section for CRP/Full Dataset parity
Move shared pages to shared level, add missing analysis pages to Full
Dataset, and relocate CMO Survey to Making of TABS.

Moves:
- CMO Survey: /results/cmo-survey -> /making-of-tabs/cmo-survey
  (redirect left at old URL)
- Statistics Glossary: /results/crp-2026/glossary -> /results/glossary
  (redirect left at old URL, CRP pages updated to link to shared glossary)

New pages:
- /results/factor-analysis (Full Dataset barrier factor structure)
- /results/validation (Full Dataset instrument validation)
  Both use live-validation.json (same schema as crp-validation.json)

Navigation:
- results-series.ts: symmetric CRP 2026 and Full Dataset sections,
  glossary + reproducibility at shared level, CMO Survey removed
- making-of-tabs-series.ts: CMO Survey added
- Sitemap updated for all new/moved URLs

Pipeline:
- daily-pipeline.yml: live-validation.json now included in analysis
  artifacts and committed to src/data/ alongside sensitivity-analysis.json

Closes #1515
