+++ __tests__/app/results/results-pages.test.tsx
@@ -601,17 +601,33 @@ describe('CRP 2026 Factor Analysis Page', () => {
   })
 })
 
-describe('CRP 2026 Glossary Page', () => {
+describe('Statistics Glossary Page (shared)', () => {
   it('renders heading', async () => {
-    const { default: Page } = await import('@/app/results/crp-2026/glossary/page')
+    const { default: Page } = await import('@/app/results/glossary/page')
     render(<Page />)
     expect(screen.getByRole('heading', { name: /Statistics Glossary/i })).toBeInTheDocument()
   })
 
   it('renders at least one glossary entry', async () => {
-    const { default: Page } = await import('@/app/results/crp-2026/glossary/page')
+    const { default: Page } = await import('@/app/results/glossary/page')
     render(<Page />)
     expect(screen.getAllByText(/Cronbach/i).length).toBeGreaterThan(0)
+  })
+})
+
+describe('Full Dataset Factor Analysis Page', () => {
+  it('renders heading', async () => {
+    const { default: Page } = await import('@/app/results/factor-analysis/page')
+    render(<Page />)
+    expect(screen.getByRole('heading', { name: /Barrier Factor Structure/i })).toBeInTheDocument()
+  })
+})
+
+describe('Full Dataset Validation Page', () => {
+  it('renders heading', async () => {
+    const { default: Page } = await import('@/app/results/validation/page')
+    render(<Page />)
+    expect(screen.getByRole('heading', { name: /Instrument Validation/i })).toBeInTheDocument()
   })
 })
